% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcnpcc.R
\name{gcnpcc}
\alias{gcnpcc}
\title{gcnpcc}
\usage{
gcnpcc(A, p, h)
}
\arguments{
\item{A}{gene expression matrix of the studied genes}

\item{p}{number of genes involved in the calculation}

\item{h}{a parameter}
}
\value{
a matrix representing pairwise gene correlations and a binary adjacency matrix indicating the presence of network edges between genes
}
\description{
Calculate partial correlation coefficients for a given set of genes and return the partial correlation matrix, with edges of the gene network estimated through adaptive thresholding
}
